Refactoring of feature execution
